refactor: extract UI logic into composables
